fix: resolve #2 , View 분리
RandomListView내부에서 선언되어 Sheet로 띄우던 WordModifySheet를 WordModifyView.swift의 새로운 파일로 분리해 selectedIndex를 Binding으로 넘겨주어 해결 (최초 상태에서 @State로 선언되었던 selectedIndex가 업데이트 되지 않아 #2 가 발생한 것으로 생각중)
