+++ WordWorld/WordWorld/Model/WordModel.swift
@@ -8,27 +8,7 @@
 
 import Foundation
 
-// 단어 배열을 가지는 struct
-struct WordsHistory : Hashable {
-    var wordCount : Int?
-    var wordArray : [String]
-}
 
-//class wordArraySaver : ObservableObject {
-//    @Published var content: WordsHistory
-//    @Published var history: Array<WordsHistory>
-//
-//    init() {
-//        self.content = WordsHistory(wordCount: 0, wordArray: [])
-//        self.history = []
-//    }
-//
-//    func makeHistory(wordCount: Int, wordArray: [String]) {
-//        self.content.wordCount = wordCount
-//        self.content.wordArray = wordArray
-//        self.history.append(content)
-//    }
-//}
 
 // MainActor를 붙여야 Warning이 뜨지 않는 이유는?
 class WordLoader : ObservableObject {
